Allow the current game state to be cleared
Context
=======

According to
https://discordapp.com/developers/docs/topics/gateway#gateway-status-update
it's possible to send a `null` value in place of a `game` object when
attempting to update the game's state. However, that has the effect of
*not* updating the current state. In other words, sending `null` will
keep the current state unchanged.

Problem
=======

It's currently impossible to reset the current game state. Steps to
reproduce:

1. change the game state via `bot.game="new game"`
2. the state is changed on discord
3. clear the state out via `bot.game = nil`
4. expected: the game state clears -- what happens instead: nothing

The reason is that we want to explicitely set `game.name = null` rather
than just `game = null` in the json we send to the server.

The fix
=======

I thought about using different approaches such as forcing to send the
payload if either `game` or `url` is different from their current state
set in the instance variables (respectively `@game` and `@streamurl`).
However that doesn't work if you want to reset the state at startup
since we have no knowledge of what those attributes were before.
The current fix always sends the payload upstream. It's not that many
bytes anyway.

Also I'm a bit confused by the `url` attribute. I can't find it anywhere
on the discord docs. Perhaps it's a legacy thing? Anyway I'm leaving it
unchanged for now.