reimplement Server#default_channel for breaking API change
Discord has updated their API to remove the previous concept of the "default channel" being the channel with the same snowflake ID as the guild. It is now described as ".. the channel with the highest position that their (the client) allows them to see (read)".
This change also allows "default channels" to be deleted and hidden just like any other channel.
https://github.com/hammerandchisel/discord-api-docs/pull/329