update cached guild data from API request in Server#update_data
Because of this, methods that rely on update_data were broken (i.e., Server#afk_channel=) if they did not provide a full server hash to update the internal data with.

The updated guild object is returned from the API::Server.update request, so we can use that to reliably update the cache.