implement ServerBan, fix ban reason in ban object
This adds a data object for server ban objects, which includes the
reason field.

Additionally, the X-Audit-Log-Reason header only provides a reason *in
the audit log*, and not on the ban object itself. The reason querystring
that is provided is what reason will be shown on the ban object itself.