remove examples of client_id
This is completely optional parameter that is currently only used to
build OAuth2 bot invite URLs. It's an unnecessary extra step for
beginners to the library, and the bot will cache the client ID from
making an API request to build the invite URL if it isn't already cached from
the bot initializer.