Fix @embed_enabled caching
If the embed on the server was disabled (`false`), Server#embed_enabled?
would be `nil` because of the tertiary condition in Server#update_data
because:

nil || false || nil #=> nil

Where the third clause would be `nil` if we had only cached the server
from GUILD_CREATE, which has no `embed_*` or `widget_*` related keys.

We can do this since in any later REST call (API::Server.update) or
GUILD_UPDATE event, all keys are present.

This also renames the internal attribute to `@embed_enabled` from
`@embed`, and changed the interrogation method to match, updating the
aliases as well (so not a breaking change)