docs typo
A Historic Topic - A vs. An

Writers sometimes confuse the use of the articles a and an. We were all taught that a precedes a word starting with a consonant and that an precedes a word starting with a vowel (a, e, i, o, u, and sometimes y).

Here’s the secret to making the rule work: The rule applies to the sound of the letter beginning the word, not just the letter itself. The way we say the word will determine whether or not we use a or an. If the word begins with a vowel sound, you must use an. If it begins with a consonant sound, you must use a.

For example, the word hour begins with the consonant h. But the h is silent, so the word has a vowel sound. Hence:

an hour

The rule works the other way as well. Take the word university. It begins with the vowel u. But the u is pronounced as if it begins with the consonant y. Hence:

a university

But consider the word umbrella, also starting with u. It starts with the vowel sound uh. Hence:

an umbrella

Another vowel with a consonant sound is o. When spoken, the letter can sound as if it begins with the consonant w. Thus, we use the a:

a one-room apartment a once-famous actor

Articles with Words Beginning with ‘h,’ a or an 

The consonant giving us the most trouble is probably h. When the h begins a word and the first syllable is strongly pronounced, you should use a.

a history of Europe (accent falls on his) a hero (accent falls on he)

But when the beginning h is weakly pronounced (historic, habitual), you may use an, especially in British English.

an historic occasion (hisTORic) an habitual offender (haBITual)

But these usages are becoming increasingly old-fashioned, so you may also use a.

a historic occasion a habitual offender

Articles with Acronyms, a or an

Finally, the rule applies to acronyms as well. If you pronounce a letter as a letter and it begins with a vowel sound, you should precede it with an. The consonants with vowel sounds include f, h, l, m, n, r, s, and x.

He flew in an SST. He fired an M‑1. He attended an FDA hearing.

By the same token, if a vowel letter, with a consonant sound, is pronounced as a letter, you should use a.

He made a U‑turn.

Got it? So what is your grade?

An A? A B? Surely not an F.