Bot#parse_mention returns an Emoji with available data (#473)
This fixes cases where a message is parsed for mentions with emoji that the bot does not have access to, and thus won't have cached, so Bot#parse_mention just returns nil on the cache miss. The emoji markdown itself still contains data we can return to the user (name, ID, whether its animated), we just won't know where it came from.