Implement CommandAlias (#524)
This adds CommandAlias as a pass-through to another command as an
alternative of "duplicating" commands with passing an array of symbols
as a name. Aliases will be shown in the help command, and asking for
help on a command alias will defer to the original command's help.

Arrays as a command name argument is marked as deprecated and
will be removed in a later release.