Sort roles by position in permissions calculation
The crux of this bug is based on the fact that the internal order of
roles is undefined: they are in whatever order Discord sends them to us.
This causes our reduce to be unstable based on this position, because of
the following, where:

- each element is a role
- "true" represents permissions field being checked
- pN is the roles position

    [false, true, false].reduce { |a, e| e } # => false
     p1     p3^^  p2
    [false, false, true].reduce { |a, e| e } # => true
     p1     p2     p3^^

Sorting during our reduction ensures a set permission by a higher role
is not masked by a lower one.